requirement.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==3.2.25
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.0
importlib-metadata==6.7.0
Markdown==3.4.4
PyJWT==2.8.0
pytz==2024.1
sqlparse==0.4.4
typing-extensions==4.7.1
zipp==3.15.0
